add ScatterChart, refactor LineChart to inherit from scatter
