+++ packages/core/demo/index.html
@@ -217,6 +217,17 @@
 			Promise
 		</button>
 	</div>
+
+	<div class="demo-chart-holder has-actions" id="classy-scatter-chart-holder"></div>
+	<div class="chart-demo-actions" id="actions-scatter" role="region" aria-label="Scatter chart actions">
+		<button class="bx--btn bx--btn--primary" id="change-data-scatter" type="button">
+			Update Data
+		</button>
+
+		<button class="bx--btn bx--btn--primary change-data-promise" type="button" data-promise-delay="2000">
+			Promise
+		</button>
+	</div>
 </body>
 
 <script src="https://unpkg.com/carbon-components@9.8.0/scripts/carbon-components.min.js" type="text/javascript"></script>
