+++ packages/core/src/index.ts
@@ -8,6 +8,7 @@ import { DonutChart, DonutCenter } from "./donut-chart";
 import { BarChart } from "./bar-chart";
 import { LineChart } from "./line-chart";
 import { ComboChart } from "./combo-chart";
+import { ScatterChart } from "./scatter-chart";
 
 import * as Configuration from "./configuration";
 const defaultColors = Configuration.options.BASE.colors;
@@ -23,5 +24,6 @@ export {
 	DonutCenter,
 	BarChart,
 	LineChart,
-	ComboChart
+	ComboChart,
+	ScatterChart
 };
